Move revision package into internal
